requirements.txt
langchain
pandas
PyYAML
streamlit
openai
ruamel.yaml